INTERACTION/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
astunparse==1.6.3
asyncio==3.4.3
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cmake==3.31.6
coloredlogs==15.0.1
discord.py==2.5.2
distro==1.9.0
elevenlabs==1.54.0
filelock==3.17.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.3.0
gast==0.6.0
gevent==24.11.1
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
keras==3.9.0
libclang==18.1.1
llvmlite==0.44.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.1.0
namex==0.0.8
networkx==3.4.2
numba==0.61.0
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
ollama==0.4.7
onnxruntime==1.21.0
openai==1.65.5
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
propcache==0.3.0
protobuf==5.29.3
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.6.1
Pygments==2.19.1
pyserial==3.5
python-dotenv==1.0.1
python-statemachine==2.5.0
PyYAML==6.0.2
RapidFuzz==3.12.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.3
setuptools==76.0.0
silero-vad==5.1.2
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
SpeechRecognition==3.14.1
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
tf_keras==2.18.0
tiktoken==0.9.0
tokenizers==0.21.0
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
transformers==4.49.0
triton==3.2.0
typing_extensions==4.12.2
urllib3==2.3.0
websocket==0.2.1
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
whisper==1.1.10
wrapt==1.17.2
yarl==1.18.3
zope.event==5.0
zope.interface==7.2
RealtimeSTT